+++ example/metro.config.js
@@ -1,14 +1,7 @@
 const {getDefaultConfig, mergeConfig} = require('@react-native/metro-config');
 
-const fs = require('fs');
 const path = require('path');
-const exclusionList = require('metro-config/src/defaults/exclusionList');
-
-const rnwPath = fs.realpathSync(
-  path.resolve(require.resolve('react-native-windows/package.json'), '..'),
-);
-
-//
+const exclusionList = require('metro-config/private/defaults/exclusionList').default;
 
 /**
  * Metro configuration
@@ -18,19 +11,17 @@ const rnwPath = fs.realpathSync(
  */
 
 const config = {
-  //
+  watchFolders: [path.resolve(__dirname, '..')],
   resolver: {
+    enableSymlinks: true,
+    nodeModulesPaths: [path.resolve(__dirname, 'node_modules')],
     blockList: exclusionList([
-      // This stops "npx @react-native-community/cli run-windows" from causing the metro server to crash if its already running
-      new RegExp(
-        `${path.resolve(__dirname, 'windows').replace(/[/\\]/g, '/')}.*`,
-      ),
-      // This prevents "npx @react-native-community/cli run-windows" from hitting: EBUSY: resource busy or locked, open msbuild.ProjectImports.zip or other files produced by msbuild
-      new RegExp(`${rnwPath}/build/.*`),
-      new RegExp(`${rnwPath}/target/.*`),
-      /.*\.ProjectImports\.zip/,
+      new RegExp(`${path.resolve(__dirname, 'ios', 'Pods')}/.*`),
+      new RegExp(`${path.resolve(__dirname, 'ios', 'DerivedData')}/.*`),
+      new RegExp(`${path.resolve(__dirname, 'ios', 'build')}/.*`),
+      new RegExp(`${path.resolve(__dirname, 'android')}/(?:.*/)?build/.*`),
+      new RegExp(`${path.resolve(__dirname, '..', '.git')}/.*`),
     ]),
-    //
   },
   transformer: {
     getTransformOptions: async () => ({