+++ ios/StarPrinter/StarPrinterSettingMaintenanceWrapper.h
@@ -1,0 +1,7 @@
+#import "React/RCTBridgeModule.h"
+#import <Foundation/Foundation.h>
+@import StarIO10ReactNative;
+
+@interface StarPrinterSettingMaintenanceWrapper : NSObject <RCTBridgeModule>
+
+@end