+++ ios/libs/StarIO10ReactNative.xcframework/ios-arm64_arm64e/StarIO10ReactNative.framework/PrivateHeaders/CommandGeneratorWrapper.h
@@ -1,0 +1,13 @@
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface CommandGeneratorWrapper : NSObject
+
++ (NSData *)generate:(int)commandGeneratorType
+              option:(NSString *)option
+        errorMessage:(NSString * _Nullable * _Nullable)errorMessage;
+
+@end
+
+NS_ASSUME_NONNULL_END