+++ ios/libs/StarIO10ReactNative.xcframework/ios-arm64_x86_64-simulator/StarIO10ReactNative.framework/PrivateHeaders/CommandParserWrapper.h
@@ -1,0 +1,11 @@
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface CommandParserWrapper : NSObject
+
++ (NSString *)parse:(int)parserType data:(NSData *)data option:(NSString *)option;
+
+@end
+
+NS_ASSUME_NONNULL_END