+++ ios/libs/StarIO10ReactNative.xcframework/ios-arm64_x86_64-simulator/StarIO10ReactNative.framework/PrivateHeaders/SmcsProvisioning.iOS.h
@@ -1,0 +1,11 @@
+#pragma once
+
+namespace StarPrntCommandUtil
+{
+namespace SmcsProvisioning
+{
+void getValue(int environment, int index, char *data, int size, int *outputSize);
+
+void getPlatformTag(char *data, int size, int *outputSize);
+} // namespace SmcsProvisioning
+} // namespace StarPrntCommandUtil