+++ ios/libs/StarIO10ReactNative.xcframework/ios-arm64_x86_64-simulator/StarIO10ReactNative.framework/PrivateHeaders/SmcsProvisioningWrapper.h
@@ -1,0 +1,12 @@
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface SmcsProvisioningWrapper : NSObject
+
++ (NSString *)valueForEnvironment:(int)environment index:(int)index;
++ (NSString *)platformTag;
+
+@end
+
+NS_ASSUME_NONNULL_END