+++ src/MaintenanceInformationType.d.ts
@@ -1,0 +1,6 @@
+export declare enum MaintenanceInformationType {
+    ThermalHeadEnergizeCount = "ThermalHeadEnergizeCount",
+    LineFeedMotorDistanceDots = "LineFeedMotorDistanceDots",
+    CutterDriveCount = "CutterDriveCount",
+    PowerOnTimeMin = "PowerOnTimeMin"
+}