+++ src/MaintenanceInformationType.ts
@@ -1,0 +1,6 @@
+export enum MaintenanceInformationType {
+    ThermalHeadEnergizeCount = 'ThermalHeadEnergizeCount',
+    LineFeedMotorDistanceDots = 'LineFeedMotorDistanceDots',
+    CutterDriveCount = 'CutterDriveCount',
+    PowerOnTimeMin = 'PowerOnTimeMin'
+}