+++ src/StarPrinterSettingMaintenanceFactory.d.ts
@@ -1,0 +1,4 @@
+import { StarPrinterSettingMaintenance } from './StarPrinterSettingMaintenance';
+export declare class StarPrinterSettingMaintenanceFactory {
+    static create(nativeMaintenance: string, nativeStarPrinter: string | undefined): Promise<StarPrinterSettingMaintenance | undefined>;
+}